remove default state
